examples/todoapp-frontend-ui-e2etest/requirements.txt
pytest==7.4.3
pytest-bdd==7.1.0
pytest-html==4.1.1
pytest-xdist==3.5.0
allure-pytest==2.13.2
selenium==4.15.2
webdriver-manager==4.0.1
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
bcrypt==4.1.2
setuptools>=65.5.0

